move extended functionality into pxl9
