+++ gpx/makefile
@@ -1,2 +1,2 @@
 include ../rules.mak
-FLAGS+=-I../pixl
+FLAGS+=-I../pxl
