+++ math/sine.cpp
@@ -8,13 +8,7 @@ void sine(std::vector<std::pair<int, int>> &coordinates)
 	static int seed = 0;
 
 	for(int i = 0; i < 360; ++i)
-	{
-		pair<int, int> c;
-		c.first = i / 3;
-		c.second = 25 * sin(M_PI * (i + seed) / 180);
-
-		coordinates.push_back(c);
-	}
+		coordinates.push_back(make_pair(i / 3, 25 * sin(M_PI * (i + seed) / 180)));
 
 	// Update the seed
 	seed += 5;
